refactor(plugins): pluginify New and Save
