+++ src/open-scd.ts
@@ -4,11 +4,8 @@ import {
   html,
   LitElement,
   property,
-  query,
   TemplateResult,
 } from 'lit-element';
-import { until } from 'lit-html/directives/until.js';
-import { translate, get } from 'lit-translate';
 
 import '@material/mwc-button';
 import '@material/mwc-checkbox';
@@ -36,66 +33,25 @@ import '@material/mwc-top-app-bar-fixed';
 
 import './filtered-list.js';
 
-import { ActionDetail } from '@material/mwc-list/mwc-list-foundation';
-import { Dialog } from '@material/mwc-dialog';
-import { Drawer } from '@material/mwc-drawer';
-import { ListItemBase } from '@material/mwc-list/mwc-list-item-base';
-
-import {
-  EditorAction,
-  newOpenDocEvent,
-  newPendingStateEvent,
-  newWizardEvent,
-  Wizard,
-  WizardInput,
-} from './foundation.js';
+import { newOpenDocEvent, newPendingStateEvent } from './foundation.js';
 import { getTheme } from './themes.js';
-import { SupportedVersion } from './schemas.js';
 
-import { Editing, newEmptySCD } from './Editing.js';
+import { Editing } from './Editing.js';
+import { Hosting } from './Hosting.js';
 import { Logging } from './Logging.js';
+import { Plugging } from './Plugging.js';
 import { Setting } from './Setting.js';
-import { InstalledPlugin, Plugging, pluginIcons } from './Plugging.js';
-import { Validating } from './Validating.js';
 import { Waiting } from './Waiting.js';
 import { Wizarding } from './Wizarding.js';
 
-import { List } from '@material/mwc-list';
-interface MenuItem {
-  icon: string;
-  name: string;
-  hint?: string;
-  actionItem?: boolean;
-  action?: (event: CustomEvent<ActionDetail>) => void;
-  disabled?: () => boolean;
-  content?: () => Promise<TemplateResult>;
-}
-
-interface Triggered {
-  trigger: () => Promise<void>;
-}
-
-interface Loader {
-  load: () => Promise<void>;
-}
-
-interface Saver {
-  save: () => Promise<void>;
-}
-
-interface Validator {
-  validate: () => Promise<void>;
-}
+import { ListItem } from '@material/mwc-list/mwc-list-item';
 
 /** The `<open-scd>` custom element is the main entry point of the
  * Open Substation Configuration Designer. */
 @customElement('open-scd')
-export class OpenSCD extends Setting(
-  Wizarding(Waiting(Validating(Plugging(Editing(Logging(LitElement))))))
+export class OpenSCD extends Hosting(
+  Setting(Wizarding(Waiting(Plugging(Editing(Logging(LitElement))))))
 ) {
-  /** The currently active editor tab. */
-  @property({ type: Number })
-  activeTab = 0;
   private currentSrc = '';
   /** The current file's URL. `blob:` URLs are *revoked after parsing*! */
   @property({ type: String })
@@ -106,9 +62,6 @@ export class OpenSCD extends Setting(
     this.currentSrc = value;
     this.dispatchEvent(newPendingStateEvent(this.loadDoc(value)));
   }
-
-  @query('#menu') menuUI!: Drawer;
-  @query('#saveas') saveUI!: Dialog;
 
   /** Loads and parses an `XMLDocument` after [[`src`]] has changed. */
   private async loadDoc(src: string): Promise<void> {
@@ -123,49 +76,6 @@ export class OpenSCD extends Setting(
     if (src.startsWith('blob:')) URL.revokeObjectURL(src);
   }
 
-  private saveAs(): void {
-    this.docName =
-      this.saveUI.querySelector('mwc-textfield')?.value || this.docName;
-    this.save();
-    this.saveUI.close();
-  }
-
-  private formatXml(xml: string, tab?: string) {
-    let formatted = '',
-      indent = '';
-
-    if (!tab) tab = '\t';
-    xml.split(/>\s*</).forEach(function (node) {
-      if (node.match(/^\/\w/)) indent = indent.substring(tab!.length);
-      formatted += indent + '<' + node + '>\r\n';
-      if (node.match(/^<?\w[^>]*[^/]$/)) indent += tab;
-    });
-    return formatted.substring(1, formatted.length - 3);
-  }
-
-  private save(): void {
-    if (this.doc) {
-      const blob = new Blob(
-        [this.formatXml(new XMLSerializer().serializeToString(this.doc))],
-        {
-          type: 'application/xml',
-        }
-      );
-
-      const a = document.createElement('a');
-      a.download = this.docName;
-      a.href = URL.createObjectURL(blob);
-      a.dataset.downloadurl = ['application/xml', a.download, a.href].join(':');
-      a.style.display = 'none';
-      document.body.appendChild(a);
-      a.click();
-      document.body.removeChild(a);
-      setTimeout(function () {
-        URL.revokeObjectURL(a.href);
-      }, 1500);
-    }
-  }
-
   private handleKeyPress(e: KeyboardEvent): void {
     let handled = false;
     const ctrlAnd = (key: string) =>
@@ -175,227 +85,17 @@ export class OpenSCD extends Setting(
     if (ctrlAnd('z')) this.undo();
     if (ctrlAnd('l')) this.logUI.open ? this.logUI.close() : this.logUI.show();
     if (ctrlAnd('m')) this.menuUI.open = !this.menuUI.open;
-    if (ctrlAnd('o')) this.menuUI.querySelector('mwc-list-item')!.click();
-    if (ctrlAnd('s')) this.save();
-    if (ctrlAnd('S')) this.saveUI.show();
+    if (ctrlAnd('o'))
+      this.menuUI.querySelector<ListItem>('mwc-list-item.loader')?.click();
+    if (ctrlAnd('s'))
+      this.menuUI.querySelector<ListItem>('mwc-list-item.saver')?.click();
+    if (ctrlAnd('S'))
+      this.menuUI
+        .querySelector<ListItem>('mwc-list-item.saver + mwc-list-item.saver')
+        ?.click();
     if (ctrlAnd('P')) this.pluginUI.show();
 
     if (handled) e.preventDefault();
-  }
-
-  private createNewProject(
-    inputs: WizardInput[],
-    wizard: Element
-  ): EditorAction[] {
-    this.docName = inputs[0].value.match(/\.s[sc]d$/i)
-      ? inputs[0].value
-      : inputs[0].value + '.scd';
-    const version = <SupportedVersion>(
-      (<ListItemBase>wizard.shadowRoot!.querySelector('mwc-list')!.selected)
-        .value
-    );
-
-    this.reset();
-
-    this.doc = newEmptySCD(this.docName.slice(0, -4), version);
-
-    return [{ actions: [], title: get('menu.new'), derived: true }];
-  }
-
-  private newProjectWizard(): Wizard {
-    return [
-      {
-        title: get('menu.new'),
-        primary: {
-          icon: 'create_new_folder',
-          label: get('create'),
-          action: (inputs, wizard) => this.createNewProject(inputs, wizard),
-        },
-        content: [
-          html`<wizard-textfield
-              id="srcName"
-              label="name"
-              value="project.scd"
-              required
-              dialogInitialFocus
-            ></wizard-textfield>
-            <mwc-list activatable>
-              <mwc-radio-list-item left value="2003"
-                >Edition 1 (Schema 1.7)</mwc-radio-list-item
-              >
-              <mwc-radio-list-item left value="2007B"
-                >Edition 2 (Schema 3.1)</mwc-radio-list-item
-              >
-              <mwc-radio-list-item left selected value="2007B4"
-                >Edition 2.1 (2007B4)</mwc-radio-list-item
-              >
-            </mwc-list>`,
-        ],
-      },
-    ];
-  }
-
-  private openNewProjectWizard() {
-    this.dispatchEvent(newWizardEvent(this.newProjectWizard()));
-  }
-
-  get menu(): (MenuItem | 'divider')[] {
-    const triggered: (MenuItem | 'divider')[] = [];
-    const loaders: (MenuItem | 'divider')[] = [];
-    const savers: (MenuItem | 'divider')[] = [];
-    const validators: (MenuItem | 'divider')[] = [];
-
-    this.triggered.forEach(plugin =>
-      triggered.push({
-        icon: plugin.icon || pluginIcons['triggered'],
-        name: plugin.name,
-        action: ae => {
-          this.dispatchEvent(
-            newPendingStateEvent(
-              (<Triggered>(
-                (<unknown>(
-                  (<List>ae.target).items[ae.detail.index].lastElementChild
-                ))
-              )).trigger()
-            )
-          );
-        },
-        disabled: (): boolean => this.doc === null,
-        content: plugin.content,
-      })
-    );
-
-    this.loaders.forEach(plugin =>
-      loaders.push({
-        icon: plugin.icon || pluginIcons['loader'],
-        name: plugin.name,
-        action: ae => {
-          this.dispatchEvent(
-            newPendingStateEvent(
-              (<Loader>(
-                (<unknown>(
-                  (<List>ae.target).items[ae.detail.index].lastElementChild
-                ))
-              )).load()
-            )
-          );
-        },
-        disabled: (): boolean => false,
-        content: plugin.content,
-      })
-    );
-
-    this.savers.forEach(plugin =>
-      savers.push({
-        icon: plugin.icon || pluginIcons['saver'],
-        name: plugin.name,
-        action: ae => {
-          this.dispatchEvent(
-            newPendingStateEvent(
-              (<Saver>(
-                (<unknown>(
-                  (<List>ae.target).items[ae.detail.index].lastElementChild
-                ))
-              )).save()
-            )
-          );
-        },
-        disabled: (): boolean => this.doc === null,
-        content: plugin.content,
-      })
-    );
-
-    this.validators.forEach(plugin =>
-      validators.push({
-        icon: plugin.icon || pluginIcons['validator'],
-        name: plugin.name,
-        action: ae => {
-          this.dispatchEvent(
-            newPendingStateEvent(
-              (<Validator>(
-                (<unknown>(
-                  (<List>ae.target).items[ae.detail.index].lastElementChild
-                ))
-              )).validate()
-            )
-          );
-        },
-        disabled: (): boolean => this.doc === null,
-        content: plugin.content,
-      })
-    );
-
-    if (triggered.length > 0) triggered.push('divider');
-
-    return [
-      'divider',
-      ...loaders,
-      ...savers,
-      {
-        icon: 'create_new_folder',
-        name: 'menu.new',
-        action: (): void => this.openNewProjectWizard(),
-      },
-      {
-        icon: 'save_alt',
-        name: 'save',
-        action: (): void => this.save(),
-        disabled: (): boolean => this.doc === null,
-      },
-      {
-        icon: 'save',
-        name: 'saveAs',
-        action: (): void => this.saveUI.show(),
-        disabled: (): boolean => this.doc === null,
-      },
-      'divider',
-      {
-        icon: 'undo',
-        name: 'undo',
-        actionItem: true,
-        action: this.undo,
-        disabled: (): boolean => !this.canUndo,
-      },
-      {
-        icon: 'redo',
-        name: 'redo',
-        actionItem: true,
-        action: this.redo,
-        disabled: (): boolean => !this.canRedo,
-      },
-      /* {
-        icon: 'rule_folder',
-        name: 'menu.validate',
-        action: () =>
-          this.doc
-            ? this.dispatchEvent(
-                newPendingStateEvent(
-                  this.validate(this.doc, { fileName: this.docName })
-                )
-              )
-            : null,
-        disabled: (): boolean => this.doc === null,
-      }, */
-      ...validators,
-      {
-        icon: 'rule',
-        name: 'menu.viewLog',
-        actionItem: true,
-        action: (): void => this.logUI.show(),
-      },
-      'divider',
-      ...triggered,
-      {
-        icon: 'settings',
-        name: 'settings.name',
-        action: (): void => this.settingsUI.show(),
-      },
-      {
-        icon: 'extension',
-        name: 'plugins.heading',
-        action: (): void => this.pluginUI.show(),
-      },
-    ];
   }
 
   constructor() {
@@ -405,131 +105,8 @@ export class OpenSCD extends Setting(
     document.onkeydown = this.handleKeyPress;
   }
 
-  renderMenuItem(me: MenuItem | 'divider'): TemplateResult {
-    if (me === 'divider')
-      return html`<li divider padded role="separator"></li>`;
-    return html`
-      <mwc-list-item
-        iconid="${me.icon}"
-        graphic="icon"
-        .disabled=${me.disabled?.() || !me.action}
-        ><mwc-icon slot="graphic">${me.icon}</mwc-icon>
-        <span>${translate(me.name)}</span>
-        ${me.hint
-          ? html`<span slot="secondary"><tt>${me.hint}</tt></span>`
-          : ''}
-        ${me.content
-          ? until(
-              me.content(),
-              html`<mwc-linear-progress indeterminate></mwc-linear-progress>`
-            )
-          : ''}
-      </mwc-list-item>
-    `;
-  }
-
-  renderActionItem(me: MenuItem | 'divider'): TemplateResult {
-    if (me !== 'divider' && me.actionItem)
-      return html`<mwc-icon-button
-        slot="actionItems"
-        icon="${me.icon}"
-        label="${me.name}"
-        ?disabled=${me.disabled?.() || !me.action}
-        @click=${me.action}
-      ></mwc-icon-button>`;
-    else return html``;
-  }
-
-  renderEditorTab({ name, icon }: InstalledPlugin): TemplateResult {
-    return html`<mwc-tab label=${translate(name)} icon=${icon || 'edit'}>
-    </mwc-tab>`;
-  }
-
   render(): TemplateResult {
-    return html`
-      <mwc-drawer class="mdc-theme--surface" hasheader type="modal" id="menu">
-        <span slot="title">${translate('menu.name')}</span>
-        ${this.docName
-          ? html`<span slot="subtitle">${this.docName}</span>`
-          : ''}
-        <mwc-list
-          wrapFocus
-          @action=${(ae: CustomEvent<ActionDetail>) =>
-            (<MenuItem>(
-              this.menu.filter(item => item !== 'divider')[ae.detail.index]
-            ))?.action?.(ae)}
-        >
-          ${this.menu.map(this.renderMenuItem)}
-        </mwc-list>
-
-        <mwc-top-app-bar-fixed slot="appContent">
-          <mwc-icon-button
-            icon="menu"
-            label="Menu"
-            slot="navigationIcon"
-            @click=${() => (this.menuUI.open = true)}
-          ></mwc-icon-button>
-          <div slot="title" id="title">${this.docName}</div>
-          ${this.menu.map(this.renderActionItem)}
-          ${this.doc
-            ? html`<mwc-tab-bar
-                @MDCTabBar:activated=${(e: CustomEvent) =>
-                  (this.activeTab = e.detail.index)}
-              >
-                ${this.editors.map(this.renderEditorTab)}
-              </mwc-tab-bar>`
-            : ``}
-        </mwc-top-app-bar-fixed>
-      </mwc-drawer>
-
-      <mwc-dialog heading="${translate('saveAs')}" id="saveas">
-        <mwc-textfield
-          dialogInitialFocus
-          label="${translate('filename')}"
-          value="${this.docName}"
-        >
-        </mwc-textfield>
-        <mwc-button
-          @click=${() => this.saveAs()}
-          icon="save"
-          slot="primaryAction"
-        >
-          ${translate('save')}
-        </mwc-button>
-        <mwc-button
-          dialogAction="cancel"
-          style="--mdc-theme-primary: var(--mdc-theme-error)"
-          slot="secondaryAction"
-        >
-          ${translate('cancel')}
-        </mwc-button>
-      </mwc-dialog>
-
-      ${this.doc
-        ? until(
-            this.editors[this.activeTab] &&
-              this.editors[this.activeTab].content(),
-            html`<mwc-linear-progress indeterminate></mwc-linear-progress>`
-          )
-        : html`<div class="landing">
-            <mwc-icon-button
-              class="landing_icon"
-              icon="create_new_folder"
-              @click=${() => this.openNewProjectWizard()}
-            >
-              <div class="landing_label">${translate('menu.new')}</div>
-            </mwc-icon-button>
-            <mwc-icon-button
-              class="landing_icon"
-              icon="folder_open"
-              @click=${() =>
-                this.menuUI.querySelector('mwc-list-item')!.click()}
-            >
-              <div class="landing_label">${translate('menu.open')}</div>
-            </mwc-icon-button>
-          </div>`}
-      ${super.render()} ${getTheme(this.settings.theme)}
-    `;
+    return html` ${super.render()} ${getTheme(this.settings.theme)} `;
   }
 
   static styles = css`
